add e2e for loadpoint and grid meter
