+++ tests/config-pv.spec.ts
@@ -1,6 +1,5 @@
 import { test, expect } from "@playwright/test";
 import { start, stop, restart, baseUrl } from "./evcc";
-import { startSimulator, stopSimulator, simulatorHost } from "./simulator";
 import { enableExperimental, expectModalHidden, expectModalVisible } from "./utils";
 
 const CONFIG_GRID_ONLY = "config-grid-only.evcc.yaml";
@@ -70,45 +69,5 @@ test.describe("pv meter", async () => {
     await restart(CONFIG_GRID_ONLY);
     await page.reload();
     await expect(page.getByTestId("pv")).toHaveCount(0);
-  });
-
-  test("remove broken pv meter", async ({ page }) => {
-    // setup test data for mock openems api
-    await startSimulator();
-
-    await page.goto("/#/config");
-    await enableExperimental(page, false);
-
-    // create meter
-    await page.getByRole("button", { name: "Add solar or battery" }).click();
-    const meterModal = page.getByTestId("meter-modal");
-    await meterModal.getByRole("button", { name: "Add solar meter" }).click();
-    await meterModal.getByLabel("Title").fill("North Roof");
-    await meterModal.getByLabel("Manufacturer").selectOption("shelly-1pm");
-    await meterModal.getByLabel("IP address or hostname").fill(simulatorHost());
-    await meterModal.getByRole("button", { name: "Validate & save" }).click();
-    await expectModalHidden(meterModal);
-    await expect(page.getByTestId("pv")).toBeVisible();
-    await expect(page.getByTestId("pv")).toContainText("North Roof");
-
-    // break meter
-    await stopSimulator();
-    await restart(CONFIG_GRID_ONLY);
-    await page.reload();
-
-    // remove meter
-    await expect(page.getByTestId("fatal-error")).toBeVisible();
-    await expect(page.getByTestId("pv")).toBeVisible();
-    await page.getByTestId("pv").getByRole("button", { name: "edit" }).click();
-    await expectModalVisible(meterModal);
-    await meterModal.getByRole("button", { name: "Delete" }).click();
-    await expectModalHidden(meterModal);
-    await expect(page.getByTestId("pv")).toHaveCount(0);
-
-    // restart and check again
-    await restart(CONFIG_GRID_ONLY);
-    await page.reload();
-    await expect(page.getByTestId("pv")).toHaveCount(0);
-    await expect(page.getByTestId("fatal-error")).not.toBeVisible();
   });
 });
